% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_NEW_tasks_Github}
\alias{check_NEW_tasks_Github}
\title{check_NEW_tasks_Github}
\usage{
check_NEW_tasks_Github()
}
\value{
Prints a CLI message
}
\description{
Check if we have new tasks in jsPsychMaker/canonical_protocol/tasks Github compared to the jsPsychMaker package (jsPsychMaker/inst/tamplates/tasks.zip)
}
\examples{
check_NEW_tasks_Github()
}
