% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_progress_pid}
\alias{check_progress_pid}
\title{check_progress_pid}
\usage{
check_progress_pid(pid = 3)
}
\arguments{
\item{pid}{protocol id}
}
\description{
Get list of files in the .data/ folder for a project
}
