% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trialids.R
\name{check_trialids}
\alias{check_trialids}
\title{check_trialids}
\usage{
check_trialids(folder_protocol, show_messages = TRUE)
}
\arguments{
\item{folder_protocol}{Folder where the protocol is}

\item{show_messages}{TRUE/FALSE}
}
\value{
messages about trialid's health
}
\description{
Checks that trialid's of an experiment in a folder follow the standard expected rules
}
