% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{copy_canonical_clean}
\alias{copy_canonical_clean}
\title{copy_canonical_clean}
\usage{
copy_canonical_clean(destination_folder)
}
\arguments{
\item{destination_folder}{Where to copy canonical_protocol_clean}
}
\description{
copy_canonical_clean
}
