% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_example_tasks.R
\name{copy_example_tasks}
\alias{copy_example_tasks}
\title{copy_example_tasks}
\usage{
copy_example_tasks(destination_folder)
}
\arguments{
\item{destination_folder}{Destination folder for the examples}
}
\description{
copy_example_tasks
}
