% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_instructions.R
\name{create_instructions}
\alias{create_instructions}
\title{create_instructions}
\usage{
create_instructions(INSTRUCTIONS, task_name, jsPsych_version = 6)
}
\arguments{
\item{INSTRUCTIONS}{HTML files or character vector}

\item{task_name}{Name of task}

\item{jsPsych_version}{By default jsPsych6. Can also be 7 for jsPsych7}
}
\value{
Returns a string with the instructions for a task
}
\description{
create_instructions
}
