% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_protocol}
\alias{create_protocol}
\title{create_protocol}
\usage{
create_protocol(
  folder_tasks = NULL,
  canonical_tasks = NULL,
  folder_output = "~/Downloads/new_protocol",
  launch_browser = FALSE,
  piloting_task = NULL
)
}
\arguments{
\item{folder_tasks}{Add folder of the protocol}

\item{canonical_tasks}{Add already available tasks to the protocol}

\item{folder_output}{Where to create the protocol}

\item{launch_browser}{TRUE/FALSE}

\item{piloting_task}{Name of task to pilot (will be only task in config.js)}
}
\description{
create_protocol
}
