% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task.R
\name{create_task}
\alias{create_task}
\title{create_task}
\usage{
create_task(task_folder, folder_output = NULL)
}
\arguments{
\item{task_folder}{Add folder of the task}

\item{folder_output}{Where to create the task}
}
\description{
This is a description
}
