% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_tasks_from_protocol}
\alias{extract_tasks_from_protocol}
\title{extract_tasks_from_protocol}
\usage{
extract_tasks_from_protocol(folder_protocol)
}
\arguments{
\item{folder_protocol}{Add folder of the protocol}
}
\description{
Extract the list of scripts in the task folder of a protocol, and a vector
with the names of the tasks
}
