% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trialids.R
\name{find_trialids}
\alias{find_trialids}
\title{find_trialids}
\usage{
find_trialids(file_name)
}
\arguments{
\item{file_name}{TASK.js file name}
}
\value{
A tibble with trialid's
}
\description{
find_trialids
}
