% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_files_github}
\alias{list_files_github}
\title{list_files_github}
\usage{
list_files_github(
  URL = "https://api.github.com/repos/gorkang/jsPsychMaker/git/trees/main?recursive=1",
  folder = "canonical_protocol"
)
}
\arguments{
\item{URL}{api url of Github project}

\item{folder}{folder inside Github project}
}
\description{
list_files_github
}
