% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_unzip}
\alias{list_unzip}
\title{list_files_zip}
\usage{
list_unzip(
  location = "jsPsychMaker",
  zip_file = "?",
  action = "list",
  destination_folder = NULL,
  files_to_unzip = NULL,
  silent = TRUE
)
}
\arguments{
\item{location}{either a package name or a folder path}

\item{zip_file}{name of zip file}

\item{action}{unzip/list}

\item{destination_folder}{where to unzip}

\item{files_to_unzip}{files inside zip to unzip}

\item{silent}{show cli messages?}
}
\description{
list_files_zip
}
