% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{replace_tasks_config_js}
\alias{replace_tasks_config_js}
\title{replace_tasks_config_js
Uses a vector of tasks to replace one of the task blocks in config.js}
\usage{
replace_tasks_config_js(
  folder_protocol,
  tasks,
  block_tasks = "randomly_ordered_tasks_1"
)
}
\arguments{
\item{folder_protocol}{Add folder of the protocol}

\item{tasks}{String vector of tasks}

\item{block_tasks}{Where to insert the tasks randomly_ordered_tasks_1 or secuentially_ordered_tasks_1}
}
\description{
replace_tasks_config_js
Uses a vector of tasks to replace one of the task blocks in config.js
}
