% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_config_js}
\alias{update_config_js}
\title{update_config_js
Uses a vector of tasks to replace one of the task blocks in config.js}
\usage{
update_config_js(
  folder_protocol,
  tasks = NULL,
  block_tasks = "secuentially_ordered_tasks_1",
  media = NULL,
  show_messages = TRUE
)
}
\arguments{
\item{folder_protocol}{Add folder of the protocol}

\item{tasks}{String vector of tasks}

\item{block_tasks}{Where to insert the tasks randomly_ordered_tasks_1 or secuentially_ordered_tasks_1}

\item{media}{list with images, video and audios media to include in config}

\item{show_messages}{TRUE/FALSE}
}
\value{
Modifies the config.js file of a protocol
}
\description{
update_config_js
Uses a vector of tasks to replace one of the task blocks in config.js
}
\examples{
\dontrun{
update_config_js(folder_protocol = "~/Downloads/MyProtocol999", 
                tasks = NULL, 
                 block_tasks = "randomly_ordered_tasks_1", 
                 media = NULL, 
                 show_messages = TRUE)
}
}
