% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_HTML.R
\name{adapt_HTML}
\alias{adapt_HTML}
\title{adapt_HTML}
\usage{
adapt_HTML(CSVs, new_plugins = NULL, folder_output)
}
\arguments{
\item{CSVs}{CSV/XML files so we can extract names of tasks}

\item{new_plugins}{Plugins of already existing tasks added}

\item{folder_output}{Where to change the index.html file}
}
\description{
adapt_HTML
}
