% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_instructions.R
\name{create_instructions}
\alias{create_instructions}
\title{create_instructions}
\usage{
create_instructions(INSTRUCTIONS, task_name)
}
\arguments{
\item{INSTRUCTIONS}{HTML files or character vector}

\item{task_name}{Name of task}
}
\description{
create_instructions
}
