% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_items_from_file.R
\name{create_items_from_file}
\alias{create_items_from_file}
\title{create_items_from_file}
\usage{
create_items_from_file(
  file_name,
  folder_output = NULL,
  options_separator = ";",
  show_messages = FALSE
)
}
\arguments{
\item{file_name}{CSV/XLS file with info about the task}

\item{folder_output}{Where do we have the canonical_protocol_clean}

\item{options_separator}{different options are by default separated by ;}

\item{show_messages}{TRUE/FALSE}
}
\value{
A char vector with the items from a file
}
\description{
create_items_from_file
}
