% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_protocol}
\alias{create_protocol}
\title{create_protocol}
\usage{
create_protocol(
  folder_tasks = NULL,
  canonical_tasks = NULL,
  folder_output = "~/Downloads/new_protocol_999",
  force_download_media = FALSE,
  launch_browser = FALSE,
  piloting_task = NULL,
  show_messages = TRUE,
  block_tasks = "randomly_ordered_tasks_1",
  jsPsych_version = 6,
  options_separator = ";"
)
}
\arguments{
\item{folder_tasks}{Add folder of the protocol}

\item{canonical_tasks}{Add already available tasks to the protocol}

\item{folder_output}{Where to create the protocol}

\item{force_download_media}{download media even if the file already exists}

\item{launch_browser}{TRUE/FALSE}

\item{piloting_task}{Name of task to pilot (will be only task in config.js)}

\item{show_messages}{TRUE/FALSE}

\item{block_tasks}{Where to insert the tasks: randomly_ordered_tasks_1 or secuentially_ordered_tasks_1}

\item{jsPsych_version}{By default jsPsych6. Can also be 7 for jsPsych7}

\item{options_separator}{different options are by default separated by ;}
}
\value{
Creates a full protocol
}
\description{
create_protocol
}
\examples{
\dontrun{

# Create a protocol with the tasks AIM, EAR and IRI in ~/Downloads/protocol999 
# and open it in a browser
  jsPsychMaker::create_protocol(canonical_tasks = c("AIM", "EAR", "IRI"),
                                folder_output = "~/Downloads/protocol999", 
                                launch_browser = TRUE)


# Create protocol using csv/xls files:

# 1) Extract example tasks
  jsPsychMaker::copy_example_tasks(destination_folder = "~/Downloads/ExampleTasks")
  
# 2) Create protocol
  jsPsychMaker::create_protocol(folder_tasks = "~/Downloads/ExampleTasks/", 
                                folder_output = "~/Downloads/protocol999", 
                                launch_browser = TRUE)


# Create a protocol with canonical tasks and tasks from csv/xls files: 

 jsPsychMaker::create_protocol(canonical_tasks = c("AIM", "EAR", "IRI"),
                                folder_tasks = "~/Downloads/ExampleTasks/", 
                                folder_output = "~/Downloads/protocol999", 
                                launch_browser = TRUE)
}

}
