% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task.R
\name{create_task}
\alias{create_task}
\title{create_task}
\usage{
create_task(folder_task, folder_output = NULL, show_messages = FALSE)
}
\arguments{
\item{folder_task}{Add folder of the task}

\item{folder_output}{Where to create the task}

\item{show_messages}{TRUE/FALSE}
}
\value{
Creates a task
}
\description{
This is a description
}
