% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_media_for_protocol}
\alias{get_media_for_protocol}
\title{get_media_for_protocol}
\usage{
get_media_for_protocol(
  all_files_js = all_files_js,
  folder_protocol,
  force_download_media = FALSE,
  show_messages = TRUE
)
}
\arguments{
\item{all_files_js}{vector with content of all js files}

\item{folder_protocol}{where is the protocol?}

\item{force_download_media}{download even if the file already exists}

\item{show_messages}{TRUE/FALSE}
}
\value{
Downloads media files and outputs a list of files
}
\description{
get_media_for_protocol
}
\examples{
\dontrun{
  get_media_for_protocol(all_files_js = readLines("canonical_protocol/tasks/INFCONS.js"),
                        folder_protocol = "canonical_protocol/",
                        force_download_media = FALSE,
                        show_messages = TRUE)
}
}
