% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_available_tasks}
\alias{list_available_tasks}
\title{list_available_tasks}
\usage{
list_available_tasks(show_help = FALSE, jsPsych_version = 6)
}
\arguments{
\item{show_help}{Show information about where to get help?}

\item{jsPsych_version}{By default jsPsych6. Can also be 7 for jsPsych7}
}
\value{
prints a list with the available tasks
}
\description{
list_available_tasks
}
\examples{
list_available_tasks()
}
