% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_available_tasks}
\alias{list_available_tasks}
\title{list_available_tasks}
\usage{
list_available_tasks(show_help = FALSE)
}
\arguments{
\item{show_help}{Show information about where to get help?}
}
\description{
list_available_tasks
}
