% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_unzip}
\alias{list_unzip}
\title{list_files_zip}
\usage{
list_unzip(
  location = "jsPsychMaker",
  zip_file = "?",
  action = "list",
  destination_folder = NULL,
  files_to_unzip = NULL,
  silent = TRUE
)
}
\arguments{
\item{location}{either a package name or a folder path}

\item{zip_file}{name of zip file}

\item{action}{unzip/list}

\item{destination_folder}{where to unzip}

\item{files_to_unzip}{files inside zip to unzip}

\item{silent}{show cli messages?}
}
\value{
Unzips or lists files inside a zip
}
\description{
Unzips or lists files inside a zip file that can be in a package or the user filesystem
}
\examples{
\dontrun{

list_unzip(
  location = "jsPsychMaker",
  zip_file = "?",
  action = "list",
  destination_folder = NULL,
  files_to_unzip = NULL,
  silent = TRUE
)

list_unzip(
  location = "jsPsychMaker",
  zip_file = "canonical_clean_6.zip",
  action = "unzip",
  destination_folder = tempdir(),
  files_to_unzip = NULL,
  silent = FALSE
)
}

}
